Move utils into sriov package
